TO BE REWORKED: Manila Server Min/Max Microversion Check
It's necessary to check whether Server.MinMicroversion <= ManilaProvisioner.MicroversionUsed <= Server.MaxMicroversion.
The gophercloud library does not provide functions for such a check.

That's why the check is added into the Manila provisioner.

As soon as gophercloud library provides functions for such check the Manila provisioner should be reworked in such a way that it uses the gophercloud library for the check.